refactored entity components
