+++ src/app/app.module.ts
@@ -65,7 +65,7 @@ import { Angulartics2Piwik } from "angulartics2/piwik";
 import { ConfigService } from "./core/config/config.service";
 import { ViewModule } from "./core/view/view.module";
 import { DashboardModule } from "./core/dashboard/dashboard.module";
-import { EntityListModule } from "./core/entity-list/entity-list.module";
+import { EntityComponentsModule } from "./core/entity-components/entity-components.module";
 
 export function configFactory(configService: ConfigService) {
   return (): Promise<any> => configService.loadConfig();
@@ -111,7 +111,7 @@ export function configFactory(configService: ConfigService) {
     MatIconModule,
     HelpModule,
     MatNativeDateModule,
-    EntityListModule,
+    EntityComponentsModule,
     DemoDataModule.forRoot([
       ...DemoChildGenerator.provider({ count: 150 }),
       ...DemoSchoolGenerator.provider({ count: 8 }),
