+++ src/app/child-dev-project/aser/aser-component/aser.component.spec.ts
@@ -1,7 +1,5 @@
 import { async, ComponentFixture, TestBed } from "@angular/core/testing";
-
 import { AserComponent } from "./aser.component";
-import { EntitySubrecordModule } from "../../../core/entity-subrecord/entity-subrecord.module";
 import { FormsModule } from "@angular/forms";
 import { ChildrenService } from "../../children/children.service";
 import { EntityMapperService } from "../../../core/entity/entity-mapper.service";
@@ -15,6 +13,8 @@ import { Database } from "../../../core/database/database";
 import { AlertService } from "app/core/alerts/alert.service";
 import { ConfirmationDialogModule } from "../../../core/confirmation-dialog/confirmation-dialog.module";
 import { FormDialogModule } from "../../../core/form-dialog/form-dialog.module";
+import { EntityComponentsModule } from "../../../core/entity-components/entity-components.module";
+import { RouterTestingModule } from "@angular/router/testing";
 
 describe("AserComponent", () => {
   let component: AserComponent;
@@ -35,9 +35,10 @@ describe("AserComponent", () => {
       imports: [
         FormsModule,
         NoopAnimationsModule,
-        EntitySubrecordModule,
+        EntityComponentsModule,
         ConfirmationDialogModule,
         FormDialogModule,
+        RouterTestingModule,
       ],
       providers: [
         DatePipe,
