+++ src/app/child-dev-project/attendance/attendance-analysis/attendance-analysis.component.spec.ts
@@ -15,7 +15,6 @@ import { AttendanceDayBlockComponent } from "../attendance-days/attendance-day-b
 import { SchoolBlockComponent } from "../../schools/school-block/school-block.component";
 import { AttendanceDaysComponent } from "../attendance-days/attendance-days.component";
 import { FormsModule } from "@angular/forms";
-import { EntitySubrecordModule } from "../../../core/entity-subrecord/entity-subrecord.module";
 import { ChildrenService } from "../../children/children.service";
 import { EntityModule } from "../../../core/entity/entity.module";
 import { NoopAnimationsModule } from "@angular/platform-browser/animations";
@@ -24,6 +23,8 @@ import { MatNativeDateModule } from "@angular/material/core";
 import { BehaviorSubject } from "rxjs";
 import { AttendanceAnalysisComponent } from "./attendance-analysis.component";
 import { ChildBlockComponent } from "../../children/child-block-list/child-block/child-block.component";
+import { EntityComponentsModule } from "../../../core/entity-components/entity-components.module";
+import { RouterTestingModule } from "@angular/router/testing";
 
 describe("AttendanceRegisterComponent", () => {
   let component: AttendanceAnalysisComponent;
@@ -54,8 +55,9 @@ describe("AttendanceRegisterComponent", () => {
         MatNativeDateModule,
         FormsModule,
         NoopAnimationsModule,
-        EntitySubrecordModule,
+        EntityComponentsModule,
         EntityModule,
+        RouterTestingModule,
       ],
       providers: [
         {
