+++ src/app/child-dev-project/attendance/attendance-block/attendance-block.component.spec.ts
@@ -8,9 +8,10 @@ import { MatInputModule } from "@angular/material/input";
 import { MatSelectModule } from "@angular/material/select";
 import { MatTooltipModule } from "@angular/material/tooltip";
 import { FormsModule } from "@angular/forms";
-import { EntitySubrecordModule } from "../../../core/entity-subrecord/entity-subrecord.module";
 import { EntityModule } from "../../../core/entity/entity.module";
 import { AttendanceDayBlockComponent } from "../attendance-days/attendance-day-block.component";
+import { EntityComponentsModule } from "../../../core/entity-components/entity-components.module";
+import { RouterTestingModule } from "@angular/router/testing";
 
 describe("AttendanceBlockComponent", () => {
   let component: AttendanceBlockComponent;
@@ -29,8 +30,9 @@ describe("AttendanceBlockComponent", () => {
         MatSelectModule,
         MatTooltipModule,
         FormsModule,
-        EntitySubrecordModule,
+        EntityComponentsModule,
         EntityModule,
+        RouterTestingModule,
       ],
     }).compileComponents();
   }));
