+++ src/app/child-dev-project/educational-material/educational-material-component/educational-material.component.ts
@@ -1,12 +1,12 @@
 import { Component, Input, OnChanges, SimpleChanges } from "@angular/core";
 import { DatePipe } from "@angular/common";
 import { EducationalMaterial } from "../model/educational-material";
-import { ColumnDescription } from "../../../core/entity-subrecord/entity-subrecord/column-description";
 import { ChildrenService } from "../../children/children.service";
-import { ColumnDescriptionInputType } from "../../../core/entity-subrecord/entity-subrecord/column-description-input-type.enum";
 import { UntilDestroy, untilDestroyed } from "@ngneat/until-destroy";
 import { Child } from "../../children/model/child";
 import { OnInitDynamicComponent } from "../../../core/view/dynamic-components/on-init-dynamic-component.interface";
+import { ColumnDescriptionInputType } from "../../../core/entity-components/entity-subrecord/column-description-input-type.enum";
+import { ColumnDescription } from "../../../core/entity-components/entity-subrecord/column-description";
 
 @UntilDestroy()
 @Component({
