+++ src/app/child-dev-project/notes/notes-of-child/notes-of-child.component.ts
@@ -5,11 +5,11 @@ import { DatePipe } from "@angular/common";
 import { ChildrenService } from "../../children/children.service";
 import moment from "moment";
 import { SessionService } from "../../../core/session/session-service/session.service";
-import { ColumnDescription } from "../../../core/entity-subrecord/entity-subrecord/column-description";
-import { ColumnDescriptionInputType } from "../../../core/entity-subrecord/entity-subrecord/column-description-input-type.enum";
 import { UntilDestroy, untilDestroyed } from "@ngneat/until-destroy";
 import { Child } from "../../children/model/child";
 import { OnInitDynamicComponent } from "../../../core/view/dynamic-components/on-init-dynamic-component.interface";
+import { ColumnDescription } from "../../../core/entity-components/entity-subrecord/column-description";
+import { ColumnDescriptionInputType } from "../../../core/entity-components/entity-subrecord/column-description-input-type.enum";
 
 /**
  * The component that is responsible for listing the Notes that are related to a certain entity
