+++ src/app/child-dev-project/notes/notes.module.ts
@@ -34,12 +34,11 @@ import { ChildMeetingNoteAttendanceComponent } from "./note-details/child-meetin
 import { MatTabsModule } from "@angular/material/tabs";
 import { MatPaginatorModule } from "@angular/material/paginator";
 import { NotePresenceListComponent } from "./note-details/note-presence-list/note-presence-list.component";
-import { EntitySubrecordModule } from "../../core/entity-subrecord/entity-subrecord.module";
 import { ConfirmationDialogModule } from "../../core/confirmation-dialog/confirmation-dialog.module";
 import { FormDialogModule } from "../../core/form-dialog/form-dialog.module";
 import { Angulartics2Module } from "angulartics2";
 import { NoteConfigLoaderService } from "./note-config-loader/note-config-loader.service";
-import { EntityListModule } from "../../core/entity-list/entity-list.module";
+import { EntityComponentsModule } from "../../core/entity-components/entity-components.module";
 
 @NgModule({
   declarations: [
@@ -51,7 +50,6 @@ import { EntityListModule } from "../../core/entity-list/entity-list.module";
   imports: [
     CommonModule,
     FormsModule,
-    EntitySubrecordModule,
     ConfirmationDialogModule,
     FlexLayoutModule,
     RouterModule,
@@ -88,7 +86,7 @@ import { EntityListModule } from "../../core/entity-list/entity-list.module";
     MatPaginatorModule,
     FormDialogModule,
     Angulartics2Module,
-    EntityListModule,
+    EntityComponentsModule,
   ],
   providers: [NoteConfigLoaderService],
 })
