+++ src/app/core/entity-components/display-text/display-text.component.spec.ts
