+++ src/app/core/entity-components/display-tick/display-tick.component.spec.ts
