+++ src/app/core/entity-components/entity-details/entity-details.component.html
