+++ src/app/core/entity-components/entity-list/entity-list.component.html
