+++ src/app/core/entity-components/entity-subrecord/column-description-input-type.enum.ts
