+++ src/app/core/entity-components/entity-subrecord/column-description.ts
