+++ src/app/core/entity-components/entity-subrecord/entity-subrecord.component.spec.ts
@@ -1,13 +1,11 @@
 import { async, ComponentFixture, TestBed } from "@angular/core/testing";
 
 import { EntitySubrecordComponent } from "./entity-subrecord.component";
-import { EntityMapperService } from "../../entity/entity-mapper.service";
 import { MockDatabase } from "../../database/mock-database";
 import { Database } from "../../database/database";
-import { EntitySchemaService } from "../../entity/schema/entity-schema.service";
-import { AlertService } from "app/core/alerts/alert.service";
 import { ConfirmationDialogService } from "../../confirmation-dialog/confirmation-dialog.service";
-import { EntitySubrecordModule } from "../entity-subrecord.module";
+import { EntityComponentsModule } from "../entity-components.module";
+import { RouterTestingModule } from "@angular/router/testing";
 
 describe("EntitySubrecordComponent", () => {
   let component: EntitySubrecordComponent;
@@ -15,11 +13,8 @@ describe("EntitySubrecordComponent", () => {
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      imports: [EntitySubrecordModule],
+      imports: [EntityComponentsModule, RouterTestingModule],
       providers: [
-        EntityMapperService,
-        EntitySchemaService,
-        AlertService,
         { provide: Database, useClass: MockDatabase },
         {
           provide: ConfirmationDialogService,
