+++ src/app/core/entity-components/keys-pipe/keys.pipe.ts
