+++ src/app/core/ui/ui/ui.component.spec.ts
@@ -36,7 +36,6 @@ import { LatestChangesModule } from "../../latest-changes/latest-changes.module"
 import { SessionModule } from "../../session/session.module";
 import { AppConfigModule } from "../../app-config/app-config.module";
 import { NoopAnimationsModule } from "@angular/platform-browser/animations";
-import { EntitySubrecordModule } from "../../entity-subrecord/entity-subrecord.module";
 import { PrimaryActionComponent } from "../primary-action/primary-action.component";
 import { AppConfig } from "../../app-config/app-config";
 import { SessionService } from "app/core/session/session-service/session.service";
@@ -46,6 +45,7 @@ import { CookieService } from "ngx-cookie-service";
 import { SwUpdate } from "@angular/service-worker";
 import { of } from "rxjs";
 import { EntitySchemaService } from "app/core/entity/schema/entity-schema.service";
+import { EntityComponentsModule } from "../../entity-components/entity-components.module";
 
 describe("UiComponent", () => {
   let component: UiComponent;
@@ -83,7 +83,6 @@ describe("UiComponent", () => {
         MatFormFieldModule,
         NoopAnimationsModule,
         AppConfigModule,
-        EntitySubrecordModule,
         ChildrenModule,
         SchoolsModule,
         SyncStatusModule,
@@ -91,6 +90,7 @@ describe("UiComponent", () => {
         LatestChangesModule,
         SessionModule,
         FlexLayoutModule,
+        EntityComponentsModule,
       ],
       providers: [
         { provide: SessionService, useValue: mockSession },
