+++ src/app/core/view/dynamic-components-map.ts
@@ -12,11 +12,11 @@ import { NotesOfChildComponent } from "../../child-dev-project/notes/notes-of-ch
 import { HealthCheckupComponent } from "../../child-dev-project/health-checkup/health-checkup-component/health-checkup.component";
 import { EducationalMaterialComponent } from "../../child-dev-project/educational-material/educational-material-component/educational-material.component";
 import { FormComponent } from "../../child-dev-project/children/child-details/form/form.component";
-import { DisplayDateComponent } from "../entity-list/display-date/display-date.component";
-import { DisplayTextComponent } from "../entity-list/display-text/display-text.component";
+import { DisplayDateComponent } from "../entity-components/display-date/display-date.component";
+import { DisplayTextComponent } from "../entity-components/display-text/display-text.component";
 import { ListAttendanceComponent } from "../../child-dev-project/children/children-list/list-attendance/list-attendance.component";
 import { SchoolBlockWrapperComponent } from "../../child-dev-project/children/children-list/school-block-wrapper/school-block-wrapper.component";
-import { DisplayTickComponent } from "../entity-list/display-tick/display-tick.component";
+import { DisplayTickComponent } from "../entity-components/display-tick/display-tick.component";
 import { ChildrenOverviewComponent } from "../../child-dev-project/schools/children-overview/children-overview.component";
 import { ChildBlockListComponent } from "../../child-dev-project/children/child-block-list/child-block-list.component";
 import { ChildBlockComponent } from "../../child-dev-project/children/child-block-list/child-block/child-block.component";
